refactor(server): move workspace naming into domain
